fix: 修复rps组装单、出入库单bug

+++ ses-app/ses-mobile-rps/src/main/java/com/redescooter/ses/mobile/rps/dm/OpeQcScooterB.java
@@ -83,7 +83,7 @@ public class OpeQcScooterB {
     * 创建人
     */
     @ApiModelProperty(value="创建人")
-    private Integer createdBy;
+    private Long createdBy;
 
     /**
     * 创建时间
